+++ src/Ext/TechnoType/Body.h
@@ -140,6 +140,7 @@ public:
 		Valueable<bool> NotHuman_RandomDeathSequence;
 
 		Valueable<InfantryTypeClass*> DefaultDisguise;
+		Valueable<UnitTypeClass*> DefaultVehicleDisguise;
 		Valueable<bool> UseDisguiseMovementSpeed;
 
 		Nullable<int> OpenTopped_RangeBonus;
@@ -482,7 +483,8 @@ public:
 			, DestroyAnim_Random { true }
 			, NotHuman_RandomDeathSequence { false }
 
-			, DefaultDisguise {}
+			, DefaultDisguise { nullptr }
+			, DefaultVehicleDisguise { nullptr }
 			, UseDisguiseMovementSpeed {}
 
 			, OpenTopped_RangeBonus {}